Adds stx-diff to diff syntax objects
This adds stx-diff, an analogue of sexp-diff but for syntax objects.
It (tries to) preserves source locations and syntax properties.

This also generalizes the marker interface to allow procedures that transform
the node rather than adding a symbol to the tree directly, which is particularly
useful when diffing and transforming syntax.